+++ test/py/matchConstraints/adjacency/20_rc_coreInsteadOfComponent.py
@@ -1,6 +1,8 @@
+post.disableInvokeMake()
+
 # test case for the bug where the constraint is checked in the core graph
 # instead of the component graph
-rLeft = ruleGMLString("""rule [
+rLeft = Rule.fromGMLString("""rule [
 	left [
 		edge [ source 1 target 2 label "-" ]
 	]
@@ -9,16 +11,16 @@ rLeft = ruleGMLString("""rule [
 		node [ id 2 label "O" ]
 	]
 ]""")
-rRight = ruleGMLString("""rule [
+rRight = Rule.fromGMLString("""rule [
 	context [
 		node [ id 1 label "C" ]
 	]
 	constrainAdj [
-		id 1 count 0 op "="
+		id 1
 		nodeLabels [ label "O" ]
+		count 0 op "="
 	]
 ]""")
 rc = rcEvaluator(inputRules)
 exp = rc.eval(rLeft *rcSuper(enforceConstraints=True)* rRight)
 assert len(exp) > 0
-for a in exp: a.print()